microservice/requirements.txt
asgiref==3.4.1
boto3==1.20.14
botocore==1.23.14
certifi==2021.10.8
charset-normalizer==2.0.8
defusedxml==0.7.1
Django==3.2.9
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-environ==0.8.1
django-storages==1.12.3
djangorestframework==3.12.4
gunicorn==20.1.0
httpie==2.6.0
idna==3.3
jmespath==0.10.0
Pygments==2.10.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
